Simplify `f()` by differentiating calling behaviour in `add()`
Prior to this change, all logic regarding how to call the callable was
in `f()`, including both determining its type (and so whether to skip
the first argument, or whether to call the callable itself or the
un-proxied function) and determining the args and kwargs.  This made
for a fairly unweildy (if clever) function.

This change moves responsibility for determining how to call the
callable based on its type to the `add()` function (at decoration time),
leaving `f()` to determine whether the provided args and kwargs are
compatible with the `definition` of the callable, and (if so) to call
its corresponding `implementation`.