Fix synchronisation of positional argument defaults
Prior to this change, the following tests would fail:
```python
@overload
def test(x, a=1, b=2):
    return {"x": x, "a": a, "b": b}

assert test(3, 4) == {"x": 3, "a": 4, "b": 2}  # == {"x": 3, "a": 4, "b": 1}
assert test(3, a=4) == {"x": 3, "a": 4, "b": 2}  # == {"x": 3, "a": 4, "b": 1}
assert test(a=3, b=4) raises TypeError  # == {"x": 1, "a": 3, "b": 4}
```
This was because defaults were consumed in order when an argument was
missing, rather than matching the default value to the argument.

This change aligns default values for positional arguments with their
corresponding positional argument, assuring that the above assertions
hold.