Add full support for complex type checking
Prior to this change, checking of complex types (e.g. `Union[int, str]`)
was disabled by virtue of such "types" not actually being of type
`type`, and so being considered of "undefined" type.

This change uses `typing.get_args()` to extract the possible actual
types from a given potentially nested complex type expression, in a way
that can then be used as the type argument in `isinstance()`
expressions.