Support use in line with `typing.overload` for type checking
Prior to this change, mypy did not know how to properly handle the
multiple type definitions specified by functions etc. overloaded by this
module's version of the `overload` decorator: used without importing
`typing.overload`, only the last signature in a file would be used for
type checking; used in conjunction with `typing.overload`, many of the
conventions defined in PEP 484 for defining overloaded functions etc.
were violated.

This change allows a function to be defined as overloaded through
repeated use of the `@overload` decorator on methods with the same fully
qualified name.  It also provides a second decorator to
`@wrap_overloaded_as` a final definition of the function, whose
implementation is ignored, but whose metadata is used to wrap the
multiple-dispatching method, in line with the expectation of the
implementation to be defined last without `@overload` decorator as per
PEP 484.

Examples of this pattern can be seen in the test class
`TestOverloadMyPyIntegration`.